refactor: split into 3 files
