+++ packages/components/tag/src/index.ts
@@ -1,2 +1,2 @@
 export { Tag } from './Tag';
-export type { TagProps, TagType } from './Tag';
+export type { TagProps, TagType } from './types';
