+++ packages/components/tag/src/types.ts
@@ -1,0 +1,25 @@
+export type TagType =
+  | 'primary'
+  | 'primary-filled'
+  | 'positive'
+  | 'negative'
+  | 'warning'
+  | 'secondary'
+  | 'muted';
+
+export type EntityStatus =
+  | 'published'
+  | 'draft'
+  | 'archived'
+  | 'changed'
+  | 'deleted'
+  | 'new';
+
+export interface TagProps {
+  tagType?: TagType;
+  style?: React.CSSProperties;
+  className?: string;
+  testId?: string;
+  children: React.ReactNode;
+  entityStatusType?: EntityStatus;
+}
